Projects/Beam_insights_Customer_Purchase_Analysis(Fixed_Window)/requirements.txt
apache-beam[gcp]
pandas
argparse
google-cloud-pubsub
